storage: Refactor mount user handling
Finding and stopping processes that use a mount point started out as a
feature for NFS only but is now used across all of the Storage page.
Let's reflect that in the code.
